# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for syncing Notion content, querying indexed knowledge, and managing runtime settings.",
        "title": "Notion RAG API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/query": {
            "post": {
                "description": "Searches indexed Notion chunks and generates an answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Query indexed knowledge",
                "parameters": [
                    {
                        "description": "Question payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.queryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/settings": {
            "get": {
                "description": "Returns stored runtime settings metadata without exposing secret values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get runtime settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/settings.PublicValues"
                        }
                    }
                }
            },
            "put": {
                "description": "Stores runtime settings in the backend data volume",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update runtime settings",
                "parameters": [
                    {
                        "description": "Settings payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.updateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stats": {
            "get": {
                "description": "Returns indexed document and last sync stats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Index stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/sync": {
            "post": {
                "description": "Crawls Notion pages, chunks them, embeds them, and stores them locally",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Sync Notion content",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Returns the backend health state",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpapi.queryRequest": {
            "type": "object",
            "properties": {
                "question": {
                    "type": "string"
                }
            }
        },
        "httpapi.updateSettingsRequest": {
            "type": "object",
            "properties": {
                "embedding_model": {
                    "type": "string"
                },
                "gemini_api_key": {
                    "type": "string"
                },
                "generation_model": {
                    "type": "string"
                },
                "notion_root_page_ids": {
                    "type": "string"
                },
                "notion_token": {
                    "type": "string"
                },
                "notion_version": {
                    "type": "string"
                }
            }
        },
        "settings.PublicValues": {
            "type": "object",
            "properties": {
                "embedding_model": {
                    "type": "string"
                },
                "gemini_api_key_set": {
                    "type": "boolean"
                },
                "generation_model": {
                    "type": "string"
                },
                "notion_root_page_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "notion_token_set": {
                    "type": "boolean"
                },
                "notion_version": {
                    "type": "string"
                }
            }
        }
    }
}